Add support for git log --pretty=fixes
When supplying a fix for earlier commit a reference to that commit should be
included. A good format is described to include first 12 digits of the SHA
value and the one-line comment. For automatic creation of these one-liners this
change add two options to global gitconfig.

See the result by running "git log --pretty=fixes"

More information can be found at:
https://www.kernel.org/doc/Documentation/SubmittingPatches